% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cycle.R
\name{.cycle_start_dt}
\alias{.cycle_start_dt}
\title{Get the start time of a cycle (local time of the first oxygen measurement recorded)}
\usage{
.cycle_start_dt(cycle)
}
\arguments{
\item{cycle}{Data frame of cycle data as returned by read_cycle().}
}
\value{
POSIXct datetime
}
\description{
Get the start time of a cycle (local time of the first oxygen measurement recorded)
}
\keyword{internal}
