% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_exp_MO2s.R
\name{fix_exp_MO2s}
\alias{fix_exp_MO2s}
\title{Adjust MO2 values using a new fish mass}
\usage{
fix_exp_MO2s(path, chamber, new_mass)
}
\arguments{
\item{path}{Character. The directory location of the AquaResp experiment on disk.
This should be the folder created by AquaResp that contains the summary files.}

\item{chamber}{Numeric. The chamber number to correct MO2 values for.}

\item{new_mass}{Numeric. The corrected fish mass in kilograms (kg).}
}
\value{
A numeric vector of corrected MO2 values for the specified chamber.
}
\description{
This function recalculates mass-specific oxygen consumption (MO2) values from an AquaResp experiment
using a new fish mass value. Like the values calculated by AquaResp, it assumes the fish has neutral buoyancy (i.e., density = 1 g/mL),
and corrects the MO2 values based on the change in fish mass and the resulting change in effective
respirometer volume.
}
\examples{
\dontrun{
# Correct MO2 values for chamber 2 with a new fish mass of 0.345 kg and a new buoyancy of 0.91 g/
corrected_MO2s <-
   fix_exp_MO2s(path = "~/data/Resp/Experiment_1", chamber = 2, new_mass = 0.345)
}
}
