% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_exp_MO2s.R
\name{fix_exp_MO2s_2}
\alias{fix_exp_MO2s_2}
\title{Test version - Adjust the MO2 data from an AquaResp chamber with a new animal mass value and density}
\usage{
fix_exp_MO2s_2(path, chamber, new_mass, density = 1)
}
\arguments{
\item{path}{Character. The directory location of the AquaResp experiment on disk. This should be the folder created by AquaResp that contains the summary files.}

\item{chamber}{Numeric. The chamber number to correct MO2 values for. Use \code{get_chambers()} to list available chambers.}

\item{new_mass}{Numeric. The corrected fish mass in kilograms (kg).}

\item{density}{Numeric. The density of the fish in g/mL. Defaults to 1.0 (neutral buoyancy).}
}
\value{
A numeric vector of corrected MO2 values for the specified chamber.
}
\description{
This function recalculates mass-specific oxygen consumption (MO2) values from an AquaResp experiment
using a new fish mass and an optional density value. It assumes the fish is neutrally buoyant unless
a specific density (in g/mL) is provided. The function uses existing metadata and MO2 data from the
experiment directory to perform the correction.
}
\examples{
\dontrun{
# Correct MO2 values for chamber 2 with a new fish mass of 0.345 kg and density of 1.05 g/mL
corrected_MO2s <-
 fix_exp_MO2s(path = "~/data/Resp/Experiment_1", chamber = 2,
              new_mass = 0.345, density = 1.05)
}
}
