% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatheadresp-package.R
\docType{package}
\name{flatheadresp-package}
\alias{flatheadresp}
\alias{flatheadresp-package}
\title{flatheadresp: Analyse respirometry data from AquaResp}
\description{
This package contains functions that streamline working with respirometry data that is produced from the AquaResp software (i.e., used in the TAF lab for flathead metabolic rate experiments), including importing AquaResp metadata and files into R, correcting for post-hoc bodymass changes, and calculating MO2 from raw data, including coefficients of determination (R^2) to overcome the bug in the AquaResp v3.0 version where any missing O_2 values resulted in erroneous or missing R^2 values.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bwwolfe/flatheadresp}
  \item Report bugs at \url{https://github.com/bwwolfe/flatheadresp/issues}
}

}
\author{
\strong{Maintainer}: Barrett Wolfe \email{barrett.wolfe@utas.edu.au}

}
\keyword{internal}
