% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cycle.R
\name{get_cycle_min_max}
\alias{get_cycle_min_max}
\title{Get min and max PO2 values for each chamber in a cycle}
\usage{
get_cycle_min_max(cycle)
}
\arguments{
\item{cycle}{Dataframe. A cycle dataframe as returned by \code{read_cycle()}.}
}
\value{
A named numeric vector of min and max \eqn{PO_{2}} values for each chamber.
}
\description{
Extracts the minimum and maximum non-zero \eqn{PO_{2}} values for each chamber from a cycle dataframe.
}
\examples{
exp_dir_path <- system.file("extdata", "aquaresp_experiment",
                            package = "flatheadresp")
cycle <- read_cycle(1, exp_dir_path)
get_cycle_min_max(cycle)
}
