% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cycle.R
\encoding{UTF-8}
\name{get_cycle_missingness}
\alias{get_cycle_missingness}
\title{Calculate percentage of missing/zero PO2 values in a cycle}
\usage{
get_cycle_missingness(cycle)
}
\arguments{
\item{cycle}{Dataframe. A cycle dataframe as returned by \code{read_cycle()}.}
}
\value{
A named vector of proportion (0 to 1) indicating missing pO2 values in
the cycle.
}
\description{
Computes the percentage of zero values in each \eqn{PO_{2}} column of a cycle
dataframe. Values greater than zero would have an incorrect \eqn{R^{2}} as
calculated by AquaResp. Missing values are defined as pO2 less than 0 or
NA/NaN etc. 0 is the common reading when a value is not recorded in one second
e.g. if the computer is busy, -300 is recorded when a firesting probe is not
in its housing.
}
\examples{
exp_dir_path <- system.file("extdata", "aquaresp_experiment",
                            package = "flatheadresp")
cycle <- read_cycle(1, exp_dir_path)
get_cycle_missingness(cycle)
}
