% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cycle.R
\name{get_cycle_R2s}
\alias{get_cycle_R2s}
\title{Calculate correlation between PO2 and time for each chamber}
\usage{
get_cycle_R2s(cycle)
}
\arguments{
\item{cycle}{Dataframe. A cycle dataframe as returned by \code{read_cycle()}.}
}
\value{
A named numeric vector of correlation coefficients and p-values for each chamber.
}
\description{
Performs Pearson correlation (r) tests between \eqn{PO_{2}} values and Unix time for each chamber in a cycle, retyrns r ^ 2
}
\examples{
exp_dir_path <- system.file("extdata", "aquaresp_experiment",
                            package = "flatheadresp")
cycle <- read_cycle(1, exp_dir_path)
get_cycle_R2s(cycle)
}
