% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cycle.R
\name{get_cycle_starts}
\alias{get_cycle_starts}
\title{Get start times for multiple cycles}
\usage{
get_cycle_starts(cycles)
}
\arguments{
\item{cycles}{A list of cycle data frames as returned by \code{read_cycle()}.}
}
\value{
A POSIXct vector of start datetimes, named by cycle index.
}
\description{
Returns the start datetime (local time of the first oxygen measurement recorded)
for each cycle in a list of cycles.
}
\examples{
exp_dir_path <- system.file("extdata", "aquaresp_experiment",
                            package = "flatheadresp")
cycles <- lapply(1:3, read_cycle, path = exp_dir_path)
get_cycle_starts(cycles)
}
