% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cycle.R
\name{get_exp_cycle_summary}
\alias{get_exp_cycle_summary}
\title{Summarize PO2 metrics across all cycles in an experiment}
\usage{
get_exp_cycle_summary(path)
}
\arguments{
\item{path}{Character. The path to the AquaResp experiment directory.}
}
\value{
A dataframe with one row per cycle and columns for each PO2 metric for each chamber.
}
\description{
Generates a summary dataframe with PO2 statistics (min, max, correlation, missingness)
for each cycle and chamber in the experiment. The correlations are calculated with
missing values (po2 = 0) omitted to get around the bug in AquaResp.
}
\examples{
\dontrun{
get_exp_cycle_summary(path = "~/data/Resp/Experiment_1")
}
}
