% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cycle.R
\name{get_n_cycles}
\alias{get_n_cycles}
\title{Get the number of AquaResp cycles in an experiment}
\usage{
get_n_cycles(path)
}
\arguments{
\item{path}{Character. The path to the AquaResp experiment directory.}
}
\value{
Integer. The number of cycles found.
}
\description{
Counts the number of cycle files (named like \code{Cycle_<number>.txt}) in the
\code{"All slopes"} folder of an AquaResp experiment directory.
}
\examples{
exp_dir_path <- system.file("extdata", "aquaresp_experiment",
                            package = "flatheadresp")
get_n_cycles(exp_dir_path)
}
