% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cycle.R
\encoding{UTF-8}
\name{pivot_cycle_summary_long}
\alias{pivot_cycle_summary_long}
\title{Pivot output from get_exp_cycle_summary() to long format}
\usage{
pivot_cycle_summary_long(df)
}
\arguments{
\item{df}{Dataframe. Output from get_exp_cycle_summary()}
}
\value{
long format summary with columns for chamber and cycle and metrics - PO_2 minimum, PO_2 maximum, cycle correlation R^2, R^2 p value, and \% missing PO_2 data.
}
\description{
This is a helper function
}
\examples{
exp_dir_path <- system.file("extdata", "aquaresp_experiment",
                            package = "flatheadresp")
x <- get_exp_cycle_summary(path = exp_dir_path)
y <- pivot_cycle_summary_long(x)
rbind(head(y),tail(y))
}
