% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cycle.R
\name{plot_cycle_po2}
\alias{plot_cycle_po2}
\title{Plot PO2 values for a specific cycle}
\usage{
plot_cycle_po2(cycle_number, path)
}
\arguments{
\item{cycle_number}{Numeric. The cycle number to plot.}

\item{path}{Character. The path to the AquaResp experiment directory.}
}
\value{
A base-R plot of time vs PO2.
}
\description{
Generates a scatter plot of PO2 values over time for each chamber in a given cycle.
}
\examples{
\dontrun{
plot_cycle_po2(cycle_number = 1, path = "~/data/Resp/Experiment_1")
}
}
