% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_exp.R
\name{plot_exp}
\alias{plot_exp}
\title{Plot MO2 values over time for an AquaResp experiment}
\usage{
plot_exp(
  path,
  chambers = NULL,
  time_col = "TIME.UNIX",
  mo2_col = "MO2",
  chamber_col = "chamber",
  show_cycle_labels = TRUE
)
}
\arguments{
\item{path}{Character. Experiment file path on disk.}

\item{chambers}{Integer vector. Optional selection of chambers:
\itemize{
\item Positive values = include only these chambers.
\item Negative values = exclude these chambers.
\item NULL = include all chambers.
}}

\item{time_col}{Character. Column name for time values (default = "TIME.UNIX").}

\item{mo2_col}{Character. Column name for MO2 values (default = "MO2").}

\item{chamber_col}{Character. Column name for chamber identifiers (default = "chamber").}

\item{show_cycle_labels}{Logical. If TRUE (default), annotate cycle numbers at the max MO2 per cycle.}
}
\value{
A ggplot object showing MO2 vs time for selected chambers.
}
\description{
Generates a ggplot visualization of \eqn{MO_{2}} values over time for each chamber in an experiment.
Points are colored and shaped by chamber, with cycle numbers optionally annotated above the highest
MO2 in each cycle.
}
\details{
The output is a ggplot object, so additional layers can be added (see examples).
}
\examples{
library(ggplot2)
exp_dir_path <- system.file("extdata", "aquaresp_experiment",
                            package = "flatheadresp")
plot_exp(path = exp_dir_path)

# include chambers 1 and 3, change y axis range
# alternatively can use + ylim(0,1200) but this will remove data beyond these limits (e.g.
# lines connecting dots beyond this MO2 range would not appear correctly
plot_exp(path = exp_dir_path, chambers = c(1, 3)) + coord_cartesian(ylim = c(0,1200))

# exclude chamber 4 (the blank (for background resp) with extreme values)
plot_exp(path = exp_dir_path, chambers = c(-4))

# remove the cycle labels
plot_exp(path = exp_dir_path, show_cycle_labels = FALSE)
}
