% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cycle.R
\name{read_cycle}
\alias{read_cycle}
\title{Read a single AquaResp cycle file}
\usage{
read_cycle(cycle_number, path)
}
\arguments{
\item{cycle_number}{Numeric. The cycle number to read.}

\item{path}{Character. The path to the AquaResp experiment directory.}
}
\value{
A dataframe containing the data from the specified cycle
}
\description{
AquaResp stores each experimental cycle (i.e. the measurement period when MO2 is calculated as a seperate file.
This function reads the data from a specific cycle file in the "All slopes" folder of an AquaResp experiment.
}
\examples{
exp_dir_path <- system.file("extdata", "aquaresp_experiment",
                            package = "flatheadresp")
read_cycle(cycle_number = 1, path = exp_dir_path)
}
