% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cycle.R
\name{summarise_experiment}
\alias{summarise_experiment}
\title{Summarise experiment metadata}
\usage{
summarise_experiment(path)
}
\arguments{
\item{path}{Character. The path to the AquaResp experiment directory.}
}
\value{
Invisibly returns NULL. Called for its side effect of printing to the console.
}
\description{
Prints a summary of key experimental details including chambers, fish masses,
respirometer volumes, cycle timing, and environmental conditions.
}
\examples{
exp_dir_path <- system.file("extdata", "aquaresp_experiment",
                             package = "flatheadresp")
summarise_experiment(path = exp_dir_path)
}
