% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chambers.R
\name{get_chambers}
\alias{get_chambers}
\title{Get the active chambers from an AquaResp experiment directory}
\usage{
get_chambers(path)
}
\arguments{
\item{path}{Character. The directory location of the AquaResp experiment on disk.
This should be the folder created by AquaResp that contains folders like "All slopes",
"Experimental information", and "Oxygen data raw".}
}
\value{
A numeric vector of active chamber numbers.
}
\description{
This function scans the experiment directory and returns the chamber numbers
for which summary data files are available. These correspond to the active chambers
used during the AquaResp experiment.
}
\examples{
\dontrun{
get_chambers(path = "~/data/Resp/Experiment_1")
}
}
