% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cycle.R
\name{get_cycle_missingness}
\alias{get_cycle_missingness}
\title{Calculate percentage of zero PO2 values in a cycle}
\usage{
get_cycle_missingness(cycle)
}
\arguments{
\item{cycle}{Dataframe. A cycle dataframe as returned by \code{read_cycle()}.}
}
\value{
A named list of percentages (0–100) indicating missingness per chamber.
}
\description{
Computes the percentage of zero values in each PO2 column of a cycle dataframe.
Values greater than zero would have an incorrect R^2 as calculated by AquaResp.
}
\examples{
\dontrun{
cycle <- read_cycle(1, "~/data/Resp/Experiment_1")
get_cycle_missingness(cycle)
}
}
