% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cycle.R
\encoding{UTF-8}
\name{get_cycle_missingness}
\alias{get_cycle_missingness}
\title{Calculate percentage of zero PO2 values in a cycle}
\usage{
get_cycle_missingness(cycle)
}
\arguments{
\item{cycle}{Dataframe. A cycle dataframe as returned by \code{read_cycle()}.}
}
\value{
A named vector of percentages (0 to 100\%) indicating missingness per chamber.
}
\description{
Computes the percentage of zero values in each \eqn{PO_{2}} column of a cycle dataframe.
Values greater than zero would have an incorrect \eqn{R^{2}} as calculated by AquaResp.
}
\examples{
exp_dir_path <- system.file("extdata", "aquaresp_experiment",
                            package = "flatheadresp")
cycle <- read_cycle(1, exp_dir_path)
get_cycle_missingness(cycle)
}
