% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chambers.R
\name{get_exp_metadata}
\alias{get_exp_metadata}
\title{Get metadata for all chambers in an AquaResp experiment}
\usage{
get_exp_metadata(path)
}
\arguments{
\item{path}{Character. The directory location of the AquaResp experiment on disk.}
}
\value{
A dataframe where each row contains metadata for one chamber.
}
\description{
This function retrieves metadata for all active chambers in the experiment directory
and returns it as a dataframe, with one row per chamber.
}
\examples{
# Locate the example AquaResp experiment directory shipped with the package
exp_dir_path <- system.file("extdata", "aquaresp_experiment",
                            package = "flatheadresp")
print(exp_dir_path)
get_exp_metadata(path = exp_dir_path)
}
