% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chambers.R
\name{get_exp_MO2s}
\alias{get_exp_MO2s}
\title{Get \eqn{MO_{2}} values from an AquaResp experiment}
\usage{
get_exp_MO2s(path, chamber = NULL)
}
\arguments{
\item{path}{Character. The directory location of the AquaResp experiment on disk.}

\item{chamber}{Optional numeric. The chamber number(s) to retrieve \eqn{MO_{2}} values for. If omitted, all chambers are returned.}
}
\value{
A dataframe with columns for chamber number, cycle number, and \eqn{MO_{2}} values.
}
\description{
This function reads the summary data files and returns mass-specific oxygen consumption (\eqn{MO_{2}})
values for one or more chambers. If no chamber is specified, data from all chambers is returned.
}
\examples{
\dontrun{
get_exp_MO2s(path = "~/data/Resp/Experiment_1") # All chambers
get_exp_MO2s(path = "~/data/Resp/Experiment_1", chamber = 1) # Chamber 1 only
}
}
