% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cycle.R
\name{plot_cycle_po2}
\alias{plot_cycle_po2}
\title{Plot PO2 values for a specific cycle}
\usage{
plot_cycle_po2(cycle_number, path)
}
\arguments{
\item{cycle_number}{Numeric. The cycle number to plot.}

\item{path}{Character. The path to the AquaResp experiment directory.}
}
\value{
A base-R plot of time vs PO₂.
}
\description{
Generates a scatter plot of PO₂ values over time for each chamber in a given cycle.
}
\examples{
exp_dir_path <- system.file("extdata", "aquaresp_experiment",
                            package = "flatheadresp")
plot_cycle_po2(cycle_number = 1, path = exp_dir_path)
}
