% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cycle.R
\name{read_cycle}
\alias{read_cycle}
\title{Read a single AquaResp cycle file}
\usage{
read_cycle(cycle_number, path)
}
\arguments{
\item{cycle_number}{Numeric. The cycle number to read.}

\item{path}{Character. The path to the AquaResp experiment directory.}
}
\value{
A dataframe containing the data from the specified cycle.
}
\description{
AquaResp stores each experimental cycle (i.e., the measurement period when \eqn{MO_{2}} is calculated)
as a separate file. This function reads the data from a cycle file in the "All slopes" folder
of an AquaResp experiment.
}
\examples{
# Locate the example AquaResp experiment directory shipped with the package
exp_dir_path <- system.file("extdata", "aquaresp_experiment",
                            package = "flatheadresp")
# Read cycle 1 from the example experiment
read_cycle(cycle_number = 1, path = exp_dir_path)
}
